requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.1.0
pydantic-settings>=2.0.0
databricks-sdk>=0.61.0
databricks-sql-connector==4.0.2
pandas>=2.0.0
sqlmodel==0.0.24